tidying tests up
